+++ src/Networks/Blockcore.Networks.XRC/Rules/XRCHeaderVersionRule.cs
@@ -1,0 +1,16 @@
+using Blockcore.Consensus.Rules;
+using Blockcore.Features.Consensus.Rules.CommonRules;
+
+namespace Blockcore.Networks.XRC.Rules
+{
+    /// <summary>
+    /// Checks if <see cref="XRCMain"/> network block's header has a valid block version.
+    /// </summary>
+    public class XRCHeaderVersionRule : HeaderVersionRule
+    {
+        /// <inheritdoc />
+        public override void Run(RuleContext context)
+        {
+        }
+    }
+}