+++ src/Node/Blockcore.Node/NetworkSelector.cs
@@ -28,6 +28,10 @@ namespace Blockcore.Node
                     nodeSettings = new NodeSettings(networksSelector: Blockcore.Networks.Stratis.Networks.Stratis, args: args);
                     break;
 
+                case "STRAX":
+                    nodeSettings = new NodeSettings(networksSelector: Blockcore.Networks.Strax.Networks.Strax, args: args);
+                    break;
+
                 case "X42":
                     nodeSettings = new NodeSettings(networksSelector: Blockcore.Networks.x42.Networks.Networks.x42, args: args);
                     break;
@@ -58,6 +62,10 @@ namespace Blockcore.Node
 
                 case "HOME":
                     nodeSettings = new NodeSettings(networksSelector: HomeCoin.Networks.Networks.HomeCoin, args: args);
+                    break;
+
+                case "SERF":
+                    nodeSettings = new NodeSettings(networksSelector: Blockcore.Networks.SERF.Networks.SERF, args: args);
                     break;
             }