+++ src/Node/Blockcore.Node/NodeBuilder.cs
@@ -45,12 +45,14 @@ namespace Blockcore.Node
                 case "BCP":
                 case "CITY":
                 case "STRAT":
+                case "STRAX":
                 case "RUTA":
                 case "EXOS":
                 case "XDS":
                 case "XLR":
                 case "IMPLX":
                 case "HOME":
+                case "SERF":
                     nodeBuilder.UsePosConsensus().AddPowPosMining().UseColdStakingWallet();
                     break;
             }